caregiverViewModel 디코딩 영역 분리 사용하지 않는 공공데이터포털 버스 도착 예정시간 삭제
