+++ app/src/main/java/com/hansung/sherpa/sendInfo/CaregiverViewModel.kt
@@ -22,11 +22,13 @@ import java.nio.ByteBuffer
 /**
  * 보호자가 FCM을 통해 '전달받은' 정보들을 관리하는 ViewModel
  *
+ * @property receiveManager 메모리 관리를 위한 단일 ReceiveManager 생성 
  * @property coordParts 사용자가 경로 안내를 진행하는 경로 리스트
  * @property colorPart 사용자가 이동하는 경로들의 이동수단 색상 (도보는 black)
  * @property passedRoute 사용자의 경로 이동 진척도
  */
 class CaregiverViewModel : ViewModel()  {
+    private val receiveManager = ReceiveManager()
     private val _coordParts = MutableLiveData(mutableStateListOf(mutableListOf(LatLng(0.0,0.0), LatLng(0.0,0.0))))
     val coordParts: LiveData<SnapshotStateList<MutableList<LatLng>>> get() = _coordParts
 
@@ -50,29 +52,18 @@ class CaregiverViewModel : ViewModel()  {
      * ※ 이후 SpecificRouteScreen으로 화면을 전환한다.
      *
      * 해당 함수는 사용자가 경로 안내를 시작했다는 알림 받는 함수로, 해당 알림이 오면 [Sherpa 내부 서버]의 navigation.transport_route를 가져온다.
+     * 예외처리: [Sherpa 내부 서버]에서 transportRoute가 조회 완료된 경우에만 화면 이동
      */
-    fun startNavigation() {
+    fun startNavigation():Boolean {
         // 이용자가 수정한 transportRoute를 조회한다.
-        val responseJson = ReceiveManager().getTransportRoute().data!!
+        val response = receiveManager.getTransportRoute()
 
-        // 디코딩 작업 (transportRoute는 디코딩되어 전달된다.)
-        val json = responseJson
-            .substring(1, responseJson.length - 1)
-            .chunked(2)
-            .map { Integer.parseInt(it, 16).toByte() }
-            .toByteArray()
-            .toString(Charsets.UTF_8)
-
-        // transportRoute는 interface로 제작되어 있어 추가적인 타입 변환 어댑터가 필요하다.
-        val gson = GsonBuilder()
-            .registerTypeAdapter(SubPath::class.java, SubPathAdapter())
-            .create()
-
-        // 정상화된 json 객체 파싱 진행
-        Log.d("API Log", "반환: ${json}")
-        val response = gson.fromJson(json, TransportRoute::class.java)
-
-        StaticValue.transportRoute = response
+        // TODO: 예외 추가할 것 (알림 달기)
+        if(response.code == 200) {
+            StaticValue.transportRoute = receiveManager.fromTransportRouteJson(response.data!!)
+            return true
+        }
+        return false
     }
 
     /**
@@ -83,21 +74,14 @@ class CaregiverViewModel : ViewModel()  {
      */
     fun devateRoute() {
         // 사용자가 수정한 transportRoute를 조회한다.
-        val responseJson = ReceiveManager().getRoute().data!!
+        val response = receiveManager.getRoute()
 
-        // 디코딩 작업 (ReceiveResponse는 디코딩되어 전달된다.)
-        val json = responseJson
-            .substring(1, responseJson.length - 1)
-            .chunked(2)
-            .map { Integer.parseInt(it, 16).toByte() }
-            .toByteArray()
-            .toString(Charsets.UTF_8)
-
-        // 정상화된 json 객체 파싱 진행
-        Log.d("API Log", "반환: ${json}")
-        val response = Gson().fromJson(json, ReceiveRouteResponse::class.java)
-
-        _coordParts.postValue(response.coordParts)
-        _colorParts.postValue(response.colorParts)
+        // 예외처리: [Sherpa 내부 서버]에서 ReceiveRouteResponse가 조회 완료된 경우에만 화면 이동
+        // TODO: 예외 추가할 것 (알림 달기)
+        if(response.code == 200) {
+            val receiveRouteResponse = receiveManager.fromReceiveRouteResponse(response.data!!)
+            _coordParts.postValue(receiveRouteResponse.coordParts)
+            _colorParts.postValue(receiveRouteResponse.colorParts)
+        }
     }
 }
