+++ app/src/main/java/com/hansung/sherpa/sendInfo/receive/ReceiveManager.kt
@@ -2,8 +2,11 @@ package com.hansung.sherpa.sendInfo.receive
 
 import android.util.Log
 import com.google.gson.Gson
+import com.google.gson.GsonBuilder
 import com.hansung.sherpa.BuildConfig
 import com.hansung.sherpa.StaticValue
+import com.hansung.sherpa.itemsetting.SubPath
+import com.hansung.sherpa.itemsetting.SubPathAdapter
 import com.hansung.sherpa.itemsetting.TransportRoute
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -89,5 +92,38 @@ class ReceiveManager {
             }
         }
         return result?: CommonResponse(500, "에러 원인을 찾을 수 없음")
+    }
+
+    fun fromTransportRouteJson(encodingJson:String):TransportRoute {
+        // 디코딩 작업 (transportRoute는 디코딩되어 전달된다.)
+        val json = encodingJson
+            .substring(1, encodingJson.length - 1)
+            .chunked(2)
+            .map { Integer.parseInt(it, 16).toByte() }
+            .toByteArray()
+            .toString(Charsets.UTF_8)
+
+        // transportRoute는 interface로 제작되어 있어 추가적인 타입 변환 어댑터가 필요하다.
+        val gson = GsonBuilder()
+            .registerTypeAdapter(SubPath::class.java, SubPathAdapter())
+            .create()
+
+        // 정상화된 json 객체 파싱 진행
+        Log.d("API Log", "반환: ${json}")
+        return gson.fromJson(json, TransportRoute::class.java)
+    }
+
+    fun fromReceiveRouteResponse(encodingJson:String):ReceiveRouteResponse {
+        // 디코딩 작업 (ReceiveResponse는 디코딩되어 전달된다.)
+        val json = encodingJson
+            .substring(1, encodingJson.length - 1)
+            .chunked(2)
+            .map { Integer.parseInt(it, 16).toByte() }
+            .toByteArray()
+            .toString(Charsets.UTF_8)
+
+        // 정상화된 json 객체 파싱 진행
+        Log.d("API Log", "반환: ${json}")
+        return Gson().fromJson(json, ReceiveRouteResponse::class.java)
     }
 }
